extend test coverage
Removes DomHandler error handler. Judging by the htmlparser2 code, there is no way to ever experience an error
