+++ lib/parse-html.js
@@ -1,11 +1,7 @@
 var htmlparser = require('htmlparser2');
 
 var parseHTML = function parseHTML (html) {
-    var handler = new htmlparser.DomHandler(function (error, dom) {
-        if (error) {
-            throw new Error(error);
-        }
-    });
+    var handler = new htmlparser.DomHandler();
 
     var parser = new htmlparser.Parser(handler, {
         lowerCaseAttributeNames: false
